pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "context-aware-ai"
version = "1.0.0"
description = "Context-Aware AI Server for Mecanum Robot -- Jetson Orin Nano Super"
requires-python = ">=3.10"
dependencies = [
    "pyzmq==24.0.0",
    "protobuf>=4.21.6,<5.0.0",
    "grpcio==1.50.0",
    "grpcio-tools==1.50.0",
    "pyyaml==6.0",
    "h5py==3.8.0",
    "Pillow>=9.4.0",
    "fastapi>=0.110.0,<1.0.0",
    "uvicorn[standard]>=0.29.0,<1.0.0",
    "websockets>=12.0",
    "setuptools<70.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-asyncio>=0.21",
    "pytest-mock>=3.12",
    "ruff>=0.4.0",
    "httpx>=0.27.0",
]
cpu = [
    "torch==2.3.1+cpu",
    "torchvision==0.18.1+cpu",
    "numpy==1.26.0",
    "opencv-python-headless==4.8.0.74",
    "ultralytics==8.3.0",
    "supervision>=0.22.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP"]
ignore = ["E501"]
